Split jacop minizinc challenges into separate workloads. This also makes it easy to add additional challenges without having to modify the java source code.

+++ resources/build-zip-from-folder-name.sh
@@ -1,13 +1,18 @@
 #!/bin/bash
 
 if [ "$1" = "" ]; then
-    echo "Please specify folder name"
+    echo "Please specify a folder name."
     exit 1
 fi
 
 #for f in `find $1 -iname *~`; do
 #    rm "$f"
 #done
+
+if [ ! -d "$1" ]; then
+    echo "There is no folder named '$1' in the current directory."
+    exit 1
+fi
 
 rm     "$1.zip"
 zip -r "$1.zip" "$1"
